--- a/schema.sql
+++ b/schema.sql
@@ -125,18 +125,23 @@ There is a many-to-many relationship between the tables animals and vets:
 an animal can visit multiple vets and one vet can be visited by multiple animals.
 Create a "join table" called visits to handle this relationship, it should also keep track of the date of the visit. */
 
-CREATE TABLE visits(
-  animal_id INTEGER REFERENCES animals(id),
-  vet_id INTEGER REFERENCES vets(id),
-  date_of_visits DATE,
-  PRIMARY KEY(animal_id, vet_id, date_of_visits)
-);
-
 /*
 CREATE TABLE visits(
   animal_id INTEGER REFERENCES animals(id),
   vet_id INTEGER REFERENCES vets(id),
-  date_of_visits DATE
+  date_of_visit DATE,
+  PRIMARY KEY(animal_id, vet_id, date_of_visit)
 );
 */
+
+CREATE TABLE visits(
+  animal_id INTEGER REFERENCES animals(id),
+  vet_id INTEGER REFERENCES vets(id),
+  date_of_visit DATE
+);
+
+/* Vet clinic database: database performance audit */   
+
+/* -- Add an email column to your owners table */
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);