--- a/temp-pr-md
+++ b/temp-pr-md
@@ -1,66 +1,60 @@
-# Vet clinic database: add "join table" for visits
+# Vet clinic database: database performance audit
 
 ## Learning objectives
-- Understand the different types of relationships between tables.
-- Prepare complex queries that answer analytical questions.
-- Use primary key & foreign key mechanism for joining tables.
+- Understand what can impact database performance.
+
+### Estimated time: 1.5h
 
 ## Description
-In this project, we will add some **many-to-many** relationships and write more complex queries to use them to answer questions.
-
+In this project we will have a chance to optimize some slow queries in our database.
 
 ### General requirements
 
-- Make sure that we used correct Gitflow.
-- Make sure that we documented our work in a professional way].
+- Make sure that we used correct [Gitflow](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/git-github/articles/gitflow.md).
+- Make sure that we documented our work [in a professional way](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/documentation/articles/professional_repo_rules.md).
 
-### Project requirements
-- Create a table named `vets` with the following columns:
-  - `id`: integer (set it as autoincremented PRIMARY KEY)
-  - `name`: string
-  - `age`: integer
-  - `date_of_graduation`: date
-- There is a **many-to-many** relationship between the tables `species` and `vets`: a vet can specialize in multiple species, and a species can have multiple vets specialized in it. Create a "join table" called `specializations` to handle this relationship.
-- There is a **many-to-many** relationship between the tables `animals` and `vets`: an animal can visit multiple vets and one vet can be visited by multiple animals. Create a "join table" called `visits` to handle this relationship, it should also keep track of the date of the visit.
-- Insert the following data for vets:
-  - Vet William Tatcher is 45 years old and graduated Apr 23rd, 2000.
-  - Vet Maisy Smith is 26 years old and graduated Jan 17th, 2019.
-  - Vet Stephanie Mendez is 64 years old and graduated May 4th, 1981.
-  - Vet Jack Harkness is 38 years old and graduated Jun 8th, 2008.
-- Insert the following data for specialties:
-  - Vet William Tatcher is specialized in Pokemon.
-  - Vet Stephanie Mendez is specialized in Digimon and Pokemon.
-  - Vet Jack Harkness is specialized in Digimon.
-- Insert the following data for visits:
-  - Agumon visited William Tatcher on May 24th, 2020.
-  - Agumon visited Stephanie Mendez on Jul 22th, 2020.
-  - Gabumon visited Jack Harkness on Feb 2nd, 2021.
-  - Pikachu visited Maisy Smith on Jan 5th, 2020.
-  - Pikachu visited Maisy Smith on Mar 8th, 2020.
-  - Pikachu visited Maisy Smith on May 14th, 2020.
-  - Devimon visited Stephanie Mendez on May 4th, 2021.
-  - Charmander visited Jack Harkness on Feb 24th, 2021.
-  - Plantmon visited Maisy Smith on Dec 21st, 2019.
-  - Plantmon visited William Tatcher on Aug 10th, 2020.
-  - Plantmon visited Maisy Smith on Apr 7th, 2021.
-  - Squirtle visited Stephanie Mendez on Sep 29th, 2019.
-  - Angemon visited Jack Harkness on Oct 3rd, 2020.
-  - Angemon visited Jack Harkness on Nov 4th, 2020.
-  - Boarmon visited Maisy Smith on Jan 24th, 2019.
-  - Boarmon visited Maisy Smith on May 15th, 2019.
-  - Boarmon visited Maisy Smith on Feb 27th, 2020.
-  - Boarmon visited Maisy Smith on Aug 3rd, 2020.
-  - Blossom visited Stephanie Mendez on May 24th, 2020.
-  - Blossom visited William Tatcher on Jan 11th, 2021.
-- Write queries to answer the following:
-  - Who was the last animal seen by William Tatcher?
-  - How many different animals did Stephanie Mendez see?
-  - List all vets and their specialties, including vets with no specialties.
-  - List all animals that visited Stephanie Mendez between April 1st and August 30th, 2020.
-  - What animal has the most visits to vets?
-  - Who was Maisy Smith's first visit?
-  - Details for most recent visit: animal information, vet information, and date of visit.
-  - How many visits were with a vet that did not specialize in that animal's species?
-  - What specialty should Maisy Smith consider getting? Look for the species she gets the most.
+### Preparation (0.5h)
 
-@ivanmvh
+For this project we need special preparation. As the goal of this project is to solve some performance issue, first we need to introduce those issues.
+In order to do that, we will populate our database with a significant number of data.
+
+Please complete the following steps:
+
+1. Make sure that we have our database set up with the schema and data from our previous projects.
+    - If we transitioned from 1.0 and we haven't complete the entire Ruby+Databases module, _we do not have the database ready._ In that case, please use the [schema.sql](./images/schema.sql) to create tables and [data.sql](./images/data.sql) to populate tables with the initial data.
+3. Run the following query to add an extra column to the owners table:
+  ``` sql
+  -- Add an email column to our owners table
+  ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+  ```
+3. Run the following statements to add data to our database (**executing them might take a few minutes**):
+  ```sql
+
+  -- This will add 3.594.280 visits considering we have 10 animals, 4 vets, and it will use around ~87.000 timestamps (~4min approx.)
+  INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;
+
+  -- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
+  insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
+  ```
+4 Depening on our machine speed, it might be enough or not. Check that by running `explain analyze SELECT COUNT(*) FROM visits where animal_id = 4`:
+     - If we get `Execution time: X ms` and X >= 1000: that should be enough, we can continue to the project requirements.
+     - If we get `Execution time: X ms` and X < 1000: please go back to point 3. and repeat until we get a value bigger than 1000ms.
+
+### Project requirements (1h)
+
+- Use `EXPLAIN ANALYZE` on the next queries to check what is happening. Take screenshots of them - they will be necessary later.
+- Find a way to decrease the execution time of the first query. Look for hints in the previous lessons.
+- Find a way to improve execution time of the other two queries.
+- While we are making changes, check if they help by running `EXPLAIN ANALYZE`. Once we succeed, take a screenshot of the `EXPLAIN ANALYZE` result showing that time actually decreased.
+- Any changes we made to our database schema should be added to the schema.sql file and commited in a new pull request.
+- If we decide to commit any `INSERT INTO` queries - remember to add the to the data.sql file. It is important to keep only the queries that change the database structure in the schema.sql file.
+- In our pull request description include screenshots with `EXPLAIN ANALYZE` results (before and after we improve the speed) for each of the 3 problematic queries.
+
+- The following queries are taking too much time (**1 sec = 1000ms can be considered as too much time for database query**). Try them on our machine to confirm it:
+  - `SELECT COUNT(*) FROM visits where animal_id = 4;`
+
+  - `SELECT * FROM visits where vet_id = 2;`
+  
+  - `SELECT * FROM owners where email = 'owner_18327@mail.com';`
+
+