… erd-VET_CLINIC 2.pgerd
… images/7-1-1-explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 with 3.5MM reg.PNG
… images/7-1-2-explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 7MILL REG.PNG
… images/7-1-3-explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 10MILL REG.PNG
… images/7-1-4-explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 14.3 MM REG.PNG
… images/7-2-1-WITHOUT INDEX explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 28.6 MM REG.PNG
… images/7-2-2-WITH INDEX Explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 28.6 MM REG.PNG
… images/7-3-1-WITHOUT INDEX Explain analyze SELECT COUNT() FROM visits vet_id EQ 2.PNG
… images/7-3-2-WITH INDEX Explain analyze SELECT COUNT() FROM visits vet_id EQ 2.PNG
… images/7-4-1-WITHOUT INDEX Explain analyze SELECT all FROM owners where email = owner_18327@mail.com.PNG
… images/7-4-2-WITH INDEX Explain analyze SELECT all FROM owners where email = owner_18327@mail.com.PNG
… others/schema-by-ERD-tool.sql