# Vet clinic database: database performance audit

## Learning objectives
- Understand what can impact database performance.

### Estimated time: 1.5h

## Description
In this project we will have a chance to optimize some slow queries in our database.

### General requirements

- Make sure that we used correct [Gitflow](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/git-github/articles/gitflow.md).
- Make sure that we documented our work [in a professional way](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/documentation/articles/professional_repo_rules.md).

### Preparation (0.5h)

For this project we need special preparation. As the goal of this project is to solve some performance issue, first we need to introduce those issues.
In order to do that, we will populate our database with a significant number of data.

Please complete the following steps:

1. Make sure that we have our database set up with the schema and data from our previous projects.
    - If we transitioned from 1.0 and we haven't complete the entire Ruby+Databases module, _we do not have the database ready._ In that case, please use the [schema.sql](./images/schema.sql) to create tables and [data.sql](./images/data.sql) to populate tables with the initial data.
3. Run the following query to add an extra column to the owners table:
  ``` sql
  -- Add an email column to our owners table
  ALTER TABLE owners ADD COLUMN email VARCHAR(120);
  ```
3. Run the following statements to add data to our database (**executing them might take a few minutes**):
  ```sql

  -- This will add 3.594.280 visits considering we have 10 animals, 4 vets, and it will use around ~87.000 timestamps (~4min approx.)
  INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;
  -- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
  insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
  ```
4 Depening on our machine speed, it might be enough or not. Check that by running `explain analyze SELECT COUNT(*) FROM visits where animal_id = 4`:
     - If we get `Execution time: X ms` and X >= 1000: that should be enough, we can continue to the project requirements.
     - If we get `Execution time: X ms` and X < 1000: please go back to point 3. and repeat until we get a value bigger than 1000ms.
 - We needed to add 28MM reg to reach more than 1000ms time without index. (We run 8 times the insert command on visits.)

![7-2-1-WITHOUT INDEX explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 28 6 MM REG](https://user-images.githubusercontent.com/817718/223357121-f0b0bc4e-8d36-42f2-939a-a6b37892a136.PNG)

### Project requirements (1h)

- Use `EXPLAIN ANALYZE` on the next queries to check what is happening. Take screenshots of them - they will be necessary later.
- Find a way to decrease the execution time of the first query. Look for hints in the previous lessons.
- Find a way to improve execution time of the other two queries.
- While we are making changes, check if they help by running `EXPLAIN ANALYZE`. Once we succeed, take a screenshot of the `EXPLAIN ANALYZE` result showing that time actually decreased.
- Any changes we made to our database schema should be added to the schema.sql file and commited in a new pull request.
- If we decide to commit any `INSERT INTO` queries - remember to add the to the data.sql file. It is important to keep only the queries that change the database structure in the schema.sql file.
- In our pull request description include screenshots with `EXPLAIN ANALYZE` results (before and after we improve the speed) for each of the 3 problematic queries.

- The following queries are taking too much time (**1 sec = 1000ms can be considered as too much time for database query**). Try them on our machine to confirm it:
  - `SELECT COUNT(*) FROM visits where animal_id = 4;`
![7-2-1-WITHOUT INDEX explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 28 6 MM REG](https://user-images.githubusercontent.com/817718/223357855-1a02c238-8cf6-451c-9ee9-f5755f6cf1f4.PNG)
![7-2-2-WITH INDEX Explain analyze SELECT COUNT() FROM visits where animal_id EQ 4 WITH 28 6 MM REG](https://user-images.githubusercontent.com/817718/223357889-c64781c8-31d4-4b1e-894b-075276119a78.PNG)

  - `SELECT * FROM visits where vet_id = 2;`
  
![7-3-1-WITHOUT INDEX Explain analyze SELECT COUNT() FROM visits vet_id EQ 2](https://user-images.githubusercontent.com/817718/223357915-d74b8885-fd51-490e-9a68-741f4902ec2a.PNG)

![7-3-2-WITH INDEX Explain analyze SELECT COUNT() FROM visits vet_id EQ 2](https://user-images.githubusercontent.com/817718/223357968-c3c17633-b1a5-42fc-bbc5-9fb289f40f07.PNG)

  - `SELECT * FROM owners where email = 'owner_18327@mail.com';`

![7-4-1-WITHOUT INDEX Explain analyze SELECT all FROM owners where email = owner_18327@mail com](https://user-images.githubusercontent.com/817718/223357986-3e51cffa-e7b5-4651-933e-da63821057a8.PNG)

![7-4-2-WITH INDEX Explain analyze SELECT all FROM owners where email = owner_18327@mail com](https://user-images.githubusercontent.com/817718/223358015-157e4672-9798-472c-b368-c3956d5cefe5.PNG)

@ivanmvh